% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amplify_operations.R
\name{amplify_list_branches}
\alias{amplify_list_branches}
\title{Lists branches for an Amplify App}
\usage{
amplify_list_branches(appId, nextToken, maxResults)
}
\arguments{
\item{appId}{[required] Unique Id for an Amplify App.}

\item{nextToken}{Pagination token. Set to null to start listing branches from start. If a
non-null pagination token is returned in a result, then pass its value
in here to list more branches.}

\item{maxResults}{Maximum number of records to list in a single response.}
}
\description{
Lists branches for an Amplify App.
}
\section{Request syntax}{
\preformatted{amplify$list_branches(
  appId = "string",
  nextToken = "string",
  maxResults = 123
)
}
}

\keyword{internal}
