% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstsCV.R
\name{dstsCV}
\alias{dstsCV}
\title{Double seasonal exponential time series cross validation}
\usage{
dstsCV(
  train.y,
  valid.y,
  s1 = 7,
  s2 = 7 * 52,
  kfold = 4,
  par = list(alpha, beta, gamma, omega, phi)
)
}
\arguments{
\item{train.y}{A numeric vector for training.}

\item{valid.y}{A numeric vector for testing.}

\item{s1}{Period of the shorter seasonal period.}

\item{s2}{Period of the longer seasonal period.}

\item{kfold}{The number of subsets by spliting \code{valid.y}}

\item{par}{The list of parameters for Double seasonal exponential time series model}
}
\description{
The cross-validation procedure based on a rolling forecasting.
}
