% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_adj_date.R
\name{find_adj_date}
\alias{find_adj_date}
\title{Find the the date which needs to be adjusted further}
\usage{
find_adj_date(data, vars = c("date", "y"), lambda = 1.5)
}
\arguments{
\item{data}{A data frame.}

\item{vars}{The column name of the data frame, date and "distance".}

\item{lambda}{A number. See \code{\link{find_outlier}}.}
}
\value{
A list contains two data frame. One is date with fixed date, and the other one is the date with fixed weekday. In the fixed date data frame, which is called \code{fixD.dat}, there are three columns and they are the month of a year, the day of the month, and number of times the date has been treated as an outlier. In the fixed weekday data frame, which is called \code{fixWD.dat}, there are four columns and they are the month of a day, the nth week of the month, the weekday of the weekday, and the number of times such day has been treated as an outlier.
}
\description{
Find the the date which needs to be adjusted further
}
