% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outlier.R
\name{find_outlier}
\alias{find_outlier}
\title{Detect outliers}
\usage{
find_outlier(x, lambda = 1.5)
}
\arguments{
\item{x}{A numeric vector.}

\item{lambda=1.5, }{A number decribes the "distance" from the group \code{x}. The larger the lambda is, the number of outliers captured is less.}
}
\value{
A list contains a character vector names \code{is_outlier} that indicates whether it is outier ("yes", "no"), and a numeric vector named "bounds" that stores the upper cutting line and lower cutting line.
}
\description{
Outliers are detected if it is larger than Q3 + lambda * (Q3-Q1) or smaller than Q1 - lambda * (Q3-Q1), where Q3 and Q1 are the third and first quantile of \code{x}.
}
\examples{
x = rnorm(100, 0, 1)
idx = c(1, 5, 10, 21, 30)
x[idx] = x[idx] + 5
find_outlier(x, lambda = 1.5)
}
