% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_lib.R
\name{find_outliers}
\alias{find_outliers}
\title{Mark outliers and determine means and standard deviation without them}
\usage{
find_outliers(dat)
}
\arguments{
\item{dat}{}
}
\value{
a tibble
}
\description{
Mark outliers and determine means and standard deviation without them
}
\keyword{internal}
