% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_hmnc.R
\name{forecast_hmnc}
\alias{forecast_hmnc}
\title{Forecasting via Harmonic Regression Model}
\usage{
forecast_hmnc(model, par = list(season = c(7, 365.25), K = c(3, 30)), h = 30)
}
\arguments{
\item{model}{An object of class \code{forecast}, the model after applying the double seasonal harmonic regression}

\item{par}{A list contains:
\itemize{
\item season. A vector displays the seasonality of the times series. The default 7 and 365.25 are for the weekly and yearly seasonality respectively
\item K. A vector displays the number of Fourier term for weekly seasonality and yearly seasonality.
}}

\item{h}{A numeric number shows the number of periods ahead to forecast (optional)}
}
\description{
Forecasting based on the double seasonal harmonic regression
}
