% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_dist.R
\name{measure_dist}
\alias{measure_dist}
\title{Measure the distance between estimated time series and actual time series}
\usage{
measure_dist(est, truth)
}
\arguments{
\item{est}{A numeric vector of estimated time series}

\item{truth}{A numeric vector of actual time series}
}
\value{
A matrix with mean error (me), root mean square error (rmse), mean percentage error (pe), and mean absolute percentage error (mape)
}
\description{
Measure the distance between estimated time series and actual time series
}
