% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_pred.R
\name{modify_pred}
\alias{modify_pred}
\title{Modify the prediction by auto-adjusting}
\usage{
modify_pred(
  data,
  vars = c("date", "y"),
  time.window = NULL,
  par = list(lambda = 1.5, n = NULL),
  holiday = NULL,
  weekday.sep = FALSE
)
}
\arguments{
\item{data}{A data framem includes a column called \code{date}.}

\item{vars}{A charater vector represents the names of date and the numeric vector of proportion error.}

\item{time.window}{A numeric vector shows the begining date and ending date, see \code{\link{build_date}}.}

\item{par}{A list of nuisance parameters. \code{lambda=1.5}, a number decribes the "distance" between the outliers and numeric variable \code{var}. \code{n}, the threshold number of times the fixed date or the fixed weekday is treated as outliers. If \code{n = NULL}, it takes the half number of years.}

\item{holiday}{A data frame includes \code{date} and \code{name} columns. If users want to adjust more holidays, simply add the date and name of the holidays. The \code{date} should be in date format, and the \code{name} should be charater.}

\item{weekday.sep}{A logical value indicating whether modify the value separately by weekdays}
}
\value{
A list includes:
\itemize{
\item \code{h.mat} A data frame with two columns \code{date} and \code{h.hat}. \code{h.hat} is the deducating/increasing proportion of estimated time series.
\item \code{fix.date}, \code{fix.weekday}, \code{holiday}. Proprotion erorr for the past several years.
}
}
\description{
Modify the prediction by auto-adjusting
}
