% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nweek.R
\name{nweek}
\alias{nweek}
\title{Nth weekday  of the month}
\usage{
nweek(date)
}
\arguments{
\item{date}{A numeric vector represents dates. The format is year month day, for example, 20190801.}
}
\value{
The data frame with two columns, \code{date} (date formate y-m-d), \code{weekday}, and the \code{Nth} weekday of the month.
}
\description{
The given date is the Nth weekday of the month
}
\examples{
nweek(c(20190101: 20190131))
}
