% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pcr_calc_slope}
\alias{pcr_calc_slope}
\title{Recalcuate standard slope of quantity vs Ct}
\usage{
pcr_calc_slope(tidy_pcr)
}
\arguments{
\item{tidy_pcr}{a object that has been tidied by \code{tidy_pcr}}
}
\value{
a tibble with an updated \code{slope} column
}
\description{
Recalcuate standard slope of quantity vs Ct
}
