% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_control.R
\name{pcr_control}
\alias{pcr_control}
\title{(Re)calculate Delta Ct mean based on given control probe}
\usage{
pcr_control(pcr, control_probe)
}
\arguments{
\item{pcr}{A tidy \code{pcr} object}

\item{control_probe}{A probe to be used as an endogenous control (eg GAPDH)}
}
\value{
A \code{pcr} object
}
\description{
(Re)calculate Delta Ct mean based on given control probe
}
\examples{
system.file("extdata", "untidy-pcr-example.xls", package = "amplify") |>
  pcr_tidy() |>
  pcr_control("GAPDH")
}
