% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_lib.R
\name{pcr_lib_qc}
\alias{pcr_lib_qc}
\title{Create library prep quality control data}
\usage{
pcr_lib_qc(lib_calc_pcr)
}
\arguments{
\item{lib_calc_pcr}{an output from \code{pcr_lib_calc}}
}
\value{
a list
}
\description{
Create library prep quality control data
}
\details{
While the output of this function on its own is can theoretically be
used to gauge library quality, it is best used in conjunction with a
function like \code{pcr_lib_calc_report}
}
\examples{

dat_path <- system.file("extdata", "untidy-standard-curve.xlsx", package = "amplify") |>
  pcr_tidy(pad_zero = TRUE) |>
  pcr_lib_calc() |>
  pcr_lib_qc()

}
