% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_lib.R
\name{pcr_lib_qc_plot_conc}
\alias{pcr_lib_qc_plot_conc}
\title{Plot concentration of libraries across samples}
\usage{
pcr_lib_qc_plot_conc(lib_qc)
}
\arguments{
\item{lib_qc}{Output of \code{pcr_lib_qc}}
}
\value{
a \code{ggplot}
}
\description{
Plot concentration of libraries across samples
}
\examples{
system.file("extdata", "untidy-standard-curve.xlsx", package = "amplify") |>
  pcr_tidy(pad_zero = TRUE) |>
  pcr_lib_calc() |>
  pcr_lib_qc() |>
  pcr_lib_qc_plot_conc()
}
