% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_lib.R
\name{pcr_lib_qc_plot_dil}
\alias{pcr_lib_qc_plot_dil}
\title{Plot standard dilutions compared to a perfect dilution}
\usage{
pcr_lib_qc_plot_dil(lib_qc)
}
\arguments{
\item{lib_qc}{Output of \code{pcr_lib_qc}}
}
\value{
a \code{ggplot}
}
\description{
Plot standard dilutions compared to a perfect dilution
}
\details{
An optimal dilution will show blue and grey dots perfectly aligned.
A plot with blue dots consistently lagging more behind the gray dots
implies the dilutions are consistent, but less dilute than a 1:10 dilution.
Likewise, a plot with blue dots that consistently outpace the gray dots
more with each passing dot signifies consistently over-diluting the
standards.

Samples are shown as red dots.
}
\examples{
system.file("extdata", "untidy-standard-curve.xlsx", package = "amplify") |>
  pcr_tidy() |>
  pcr_lib_calc() |>
  pcr_lib_qc() |>
  pcr_lib_qc_plot_dil()
}
