% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_lib.R
\name{pcr_lib_qc_plot_outliers}
\alias{pcr_lib_qc_plot_outliers}
\title{Plot mean centered samples without putative outliers}
\usage{
pcr_lib_qc_plot_outliers(lib_qc)
}
\arguments{
\item{lib_qc}{Output of \code{pcr_lib_qc}}
}
\value{
a \code{ggplot}
}
\description{
Plot mean centered samples without putative outliers
}
\details{
A sample is deemed an outlier if, upon its removal, it is more that
3Z from the mean of the remaining. This boundary of +/-3Z is demarcated by
the shaded area. Gray samples are outliers. Samples |Z| > 10 away are
denoted by arrows (<<<) pointing in their direction as well as with their Z
}
\examples{
system.file("extdata", "untidy-standard-curve.xlsx", package = "amplify") |>
  pcr_tidy(pad_zero = TRUE) |>
  pcr_lib_calc() |>
  pcr_lib_qc() |>
  pcr_lib_qc_plot_outliers()
}
