% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_lib.R
\name{pcr_lib_qc_plot_slope}
\alias{pcr_lib_qc_plot_slope}
\title{Plot the log of library quantities vs Ct}
\usage{
pcr_lib_qc_plot_slope(lib_qc)
}
\arguments{
\item{lib_qc}{Output of \code{pcr_lib_qc}}
}
\value{
a \code{ggplot}
}
\description{
Plot the log of library quantities vs Ct
}
\details{
An optimal plot will have a slope of -3.32. This is because we
expect that a sample 10x more concentrated than another will reach the same
abundance in 3.32 doublings FASTER (that is, 3.32 fewer doubles, or Cts).
Therefore, for each 10x increase in concentration (one point left to right
on the plot) we expect a decrease in CT of 3.32. A steeper slope (more
negative) implies a poorer efficiency (more cycles are required to reach
10x than perfect doubling would imply)
}
\examples{
system.file("extdata", "untidy-standard-curve.xlsx", package = "amplify") |>
  pcr_tidy(pad_zero = TRUE) |>
  pcr_lib_calc() |>
  pcr_lib_qc() |>
  pcr_lib_qc_plot_slope()
}
