% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_lib.R
\name{pcr_lib_qc_report}
\alias{pcr_lib_qc_report}
\title{Generate visual library prep pcr quality control report}
\usage{
pcr_lib_qc_report(pcr_lib_qc, report_path = NULL)
}
\arguments{
\item{pcr_lib_qc}{output from \code{pcr_lib_qc}}

\item{report_path}{the name of the report as well as where it should be
output. If NULL, it will export to a temp directory}
}
\value{
The path to the report
}
\description{
Generate visual library prep pcr quality control report
}
\examples{
system.file("extdata", "untidy-standard-curve.xlsx", package = "amplify") |>
  pcr_tidy() |>
  pcr_lib_calc() |>
  pcr_lib_qc() |>
  pcr_lib_qc_report()
}
