% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_plan.R
\name{pcr_plan_report}
\alias{pcr_plan_report}
\title{Create a report from a PCR plan}
\usage{
pcr_plan_report(pcr_plan, file_path = NULL)
}
\arguments{
\item{pcr_plan}{output from \code{pcr_plan}}

\item{file_path}{Where the report should be written, as well as the file name. Defaults to temp file.}
}
\value{
a named list, like the output of \code{pcr_plan}, but with the output file path appended.
}
\description{
Create a report from a PCR plan
}
\examples{
dummy_rna_conc |>
  pcr_plan(n_primers = 3) |>
  pcr_plan_report()
}
