% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_plot.R
\name{pcr_plot}
\alias{pcr_plot}
\title{Plot qPCR results}
\usage{
pcr_plot(pcr)
}
\arguments{
\item{pcr}{a tidy \code{pcr} object}
}
\value{
a \code{ggplot}
}
\description{
Plot qPCR results
}
\examples{
system.file("extdata", "untidy-pcr-example.xls", package = "amplify") |>
  pcr_tidy() |>
  pcr_rq("RD1") |>
  pcr_plot()
}
