% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_rq.R
\name{pcr_rq}
\alias{pcr_rq}
\title{(Re)calculate rq for a given sample}
\usage{
pcr_rq(pcr, relative_sample)
}
\arguments{
\item{pcr}{A \code{pcr} object}

\item{relative_sample}{A sample to set others relative to (eg my_dmso_sample)}
}
\value{
A \code{tibble}
}
\description{
(Re)calculate rq for a given sample
}
\examples{
dat_path <- system.file("extdata", "untidy-pcr-example.xls", package = "amplify")

pcr_tidy(dat_path) |>
  pcr_rq("U6D1")

# Can also be run after using pcr_control:
pcr_tidy(dat_path) |>
  pcr_control("GAPDH") |>
  pcr_rq("U6D1")
}
