% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-gp.R, R/reexport-mop.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{well_data}
\alias{read_pcr}
\alias{as_pcr}
\alias{tidy_lab}
\alias{scrub}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{gp}{\code{\link[gp]{well_data}}}

  \item{mop}{\code{\link[mop]{as_pcr}}, \code{\link[mop]{read_pcr}}, \code{\link[mop]{scrub}}, \code{\link[mop]{tidy_lab}}}
}}

