% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_model_hmnc.R
\name{select_model_hmnc}
\alias{select_model_hmnc}
\title{Select the optimal Harmonic Regression Model}
\usage{
select_model_hmnc(
  train.y,
  valid.y = NULL,
  s1 = 7,
  s2 = 365.25,
  grid.search = FALSE,
  search.length = c(20, 40),
  length.out = 5
)
}
\arguments{
\item{train.y}{A numeric vector for training.}

\item{valid.y}{A numeric vector for validating If \code{grid.search = FALSE}, \code{valid.y = NULL}.}

\item{s1}{Period of the shorter seasonal period.}

\item{s2}{Period of the longer seasonal period.}

\item{grid.search}{If TRUE, a grid search is applied.}

\item{search.length}{Grid search parameter. Only used if \code{grid.search = TRUE}. It is the range of the number of the Fourier term of yearly seasonality.}

\item{length.out}{Grid search parameter. Only used if \code{grid.search = TRUE}. It is the desired length of search sequence.}
}
\value{
A list contains:
\itemize{
\item model. An object of class \code{forecast}. The model is built by combining training and testing;
\item K. A vector display the number of Fourier term for weekly seasonality and yearly seasonality.
\item cv. A data frame of searched parameters with respect to its distance measures (see \code{\link{measure_dist}})
}
}
\description{
double seasonal harmonic regression is implemented. If grid.search = TRUE, grid search is applied by searching the number of Fourier terms for yearly seasonality. \code{search.length} and \code{length.out} are for grid search. The search is based on the minimum mean absolute percentage error with parallel computing. More details about parallel computing can be found in \code{doParallel}.
}
\examples{
data(tickets)
data.ls = train_test_split(tickets, var = "date", train.window = c(20140701, 20180630), test.window = c(20180701, 20190630))
train.y = data.ls$train.dat[,2]
valid.y = data.ls$test.dat[,2]
select.ls = select_model_hmnc(train.y, valid.y, grid.search = TRUE, length.out = 5)

}
