% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_ts.R
\name{smooth_ts}
\alias{smooth_ts}
\title{Smooth time series}
\usage{
smooth_ts(
  data,
  vars = c("date", "ts"),
  par = list(lambda = 1.5, s = 7),
  week.sep = FALSE
)
}
\arguments{
\item{data}{A data frame includes dates and time series columns}

\item{vars}{A character vector with variable names with respect to date and time series}

\item{par}{A list of parameters. \code{lambda = 1.5}, the "distance" describes the outlier and the main population. \code{s = 7}, the most obvious seasonality of the time series is 7 (days).}

\item{week.sep}{A logical value indicating whether model each weekday time series separately.}
}
\value{
A list with:
\itemize{
\item ts.dat. A data frame with column name \code{is_outlier} and \code{ts}. \code{is_outlier} is a character vector that indicates whether the observation is outier ("yes", "no"). \code{ts} is a numeric vector that stores the time series after smoothing.
\item model. An object of class "ets", see \code{\link[forecast]{ets}}.
}
}
\description{
Smooth the time series by impute the outlers.
}
\details{
Use \code{find_outlier} to find the outliers and use one seasonal exponential smoothing to imput outliers
}
\examples{
data(tickets)
smooth.dat = smooth_ts(tickets, vars = c("date", "ticket"))
cbind(tickets, smooth.dat)
}
\seealso{
\code{\link{find_outlier}}
}
