% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{Training and testing set split}
\usage{
train_test_split(data, var = "date", train.window, test.window)
}
\arguments{
\item{data}{A data frame to be splitted.}

\item{var}{The column name for dates.}

\item{test.window}{A nummric vetor of testing set starting time, and ending time}

\item{trian.window}{A nummric vetor of training set starting time, and ending time}
}
\value{
A list of training set and testing set
}
\description{
Training and testing set split
}
\examples{
data(tickets)
train_test_split(tickets, var = "date", train.window = c(20140701, 20180630), test.window = c(20180701, 20190630))
}
