% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_control.R
\name{pcr_control}
\alias{pcr_control}
\alias{pcr_control.pcr}
\alias{pcr_control.data.frame}
\title{Calculate Delta Ct mean based on given control probe}
\usage{
pcr_control(x, control_probe)

\method{pcr_control}{pcr}(x, control_probe, ...)

\method{pcr_control}{data.frame}(x, control_probe, ...)
}
\arguments{
\item{x}{A \code{pcr} or \code{data.frame} object}

\item{control_probe}{A probe to be used as an endogenous control (eg GAPDH)}
}
\value{
An object with class the same as input
}
\description{
Calculate Delta Ct mean based on given control probe
}
\examples{
system.file("extdata", "untidy-pcr-example.xls", package = "amplify") |>
  read_pcr() |>
  pcr_control("GAPDH")
}
