% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_plot.R
\name{pcr_plot}
\alias{pcr_plot}
\alias{pcr_plot.pcr}
\alias{pcr_plot.data.frame}
\title{Plot qPCR results}
\usage{
pcr_plot(x, ...)

\method{pcr_plot}{pcr}(x, ...)

\method{pcr_plot}{data.frame}(x, ...)
}
\arguments{
\item{x}{a \code{pcr} object or \code{data.frame}}
}
\value{
a \code{ggplot}
}
\description{
Plot qPCR results
}
\examples{
system.file("extdata", "untidy-pcr-example.xls", package = "amplify") |>
  pcr_tidy() |>
  pcr_rq("RD1") |>
  pcr_plot()
}
