% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_rq.R
\name{pcr_rq}
\alias{pcr_rq}
\alias{pcr_rq.pcr}
\alias{pcr_rq.data.frame}
\title{Recalculate relative quantities for a given experiment}
\usage{
pcr_rq(x, relative_sample, control_probe = NULL, ...)

\method{pcr_rq}{pcr}(x, relative_sample, control_probe = NULL, ...)

\method{pcr_rq}{data.frame}(x, relative_sample, control_probe = NULL, ...)
}
\arguments{
\item{x}{A \code{pcr} or \code{data.frame}}

\item{relative_sample}{A sample to set others relative to (eg \code{my_dmso_sample})}

\item{control_probe}{Character. \code{target_name} to serve as endogenous control.}

\item{...}{Arguments passed to respective method}
}
\value{
An object of same class as \code{x}
}
\description{
Recalculate relative quantities for a given experiment
}
\examples{
dat_path <- system.file("extdata", "untidy-pcr-example.xls", package = "amplify")

pcr_tidy(dat_path) |>
  pcr_rq("U6D1")

# Can also be run after using pcr_control:
pcr_tidy(dat_path) |>
  pcr_control("GAPDH") |>
  pcr_rq("U6D1")
}
